Add InputType module and first Class
This module will house two classes which will control how the
Calculator#parse_input method behaves. I am abstracting this logic
out of that class to make it a bit easier to expand this calculator
in the future. With these comparison classes separated, the
calculator's functionality could be easily modified in the future to
allow for more operations than just basic elementary arithmetic.

It will also give the ability to further tweak how the comparisons are
made to account for any unforeseen edge cases for the input string given
to Calculator#parse_input.