Add #parse_input method to Calculator class
This method will serve as the main entry point into the application.
It will parse the given input and decide what to do with the string.
If its a valid operand, it will be added to the stack. If it is a
valid operator, it will execute an operation. Otherwise, it will
fail and display an Error to the user that their input was not valid.