Implement code to execute an operation when given operator
When a user inputs a valid operator into parse_input, it will route
to the appropriate case statement and execute the operation. If somehow
an invalid operator gets passed the case statement, it will raise an
Operation error. A successful operation will push the result of that
operation onto the stack, replacing the two operands that it used.