Add a method to print the result of sending input to the calculator
Since pushing an Operand to the stack results in a return of the
Stack contents itself, the stack is constantly outputted to the terminal
each time #parse_input is called.

However, I want to still see the output of the #parse_input method if an
Operator is inputted. The #calc_print method will check the result of
#parse_input and determine whether or not to run #puts. If the output is
an Array, we do not want to see that in the terminal. If it is just an
Integer or Float as the result of a calculation, print that with an
equals sign so the user can easily see the result of their calculation.