Refactor Regex comparison to Array comparison
This refactors the comparisons made when an operator is entered to
check an array of operators instead of using a Regex comparison. It is
less error prone and easier to understand, while having the bigger
benefit of being more explicit in its verification ability. I also
believe it is much easier to read and understand now.